Changed the AddCopyright class to UpdateCopyrightAndVersion. This class is now able to both: - update the copyright in the beggining of each Java files - replaces generic strings in java, html, xml or xsl files
The class contains generic patterns that will be looked for, as well as their replacements.

This is useful to have Versions and Copyright information at one single place in the code. The UpdateCopyrightAndVersion is to be called when a new ReleaseBranch is made.

The proactive.xml former target copyright as been replaced by update_copyright_and_version accordingly.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@4424 28e8926c-6b08-0410-baaa-805c5e19b8d6