[java] >>> Successful tests: 101 / 101
woohooo non-regression tests run successfully. 

The remaining issue was related to the lately introduced namespace for the deployment descriptors
It seems that in XPath, not having a prefix would put it in no namespace mode not in default namespace mode.
Adding a prefix for xpath attribute in all keyrefs solved the issue.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@4802 28e8926c-6b08-0410-baaa-805c5e19b8d6