added the "Synchronous" method annotation to allow the user to tell proactive that a method is expected to be called synchronously.
The "Synchronous method call" warning message will therefore be diplayed if this annotation is not present on a method or if this method is toString or hashCode.

This annotation is not meant to change the method behaviour yet, e.g. a method that would be reified asynchronously will still be reified asynchronously if the annotation is present. 



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@6388 28e8926c-6b08-0410-baaa-805c5e19b8d6