Modify the way of using the database. Remove db.cfg file in database package Rename db.cfg into scheduler_db.cfg in the "script/unix/scheduler" directory
Database has to be created using "createDatabase.sh" and the "scheduler_db.cfg" configuration file.
Scheduler has to be start with a configuration file.
This file can be set using the "-c conf_file" option when starting the scheduler. Default file is the "scheduler_db.cfg" file into the "script/unix/scheduler" directory

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@8705 28e8926c-6b08-0410-baaa-805c5e19b8d6