Commit of the NAS Parallel Benchmarks.
These are a set of 5 benchmarks designed by NASA which are representative of common Computational Fluid Dynamics (CFD) computations:
CG: is a strong test for small communication (~10,000 messages of few bytes per worker)
MG: is a strong test for mid-size communication (~2,500 messages (2KB-512KB) per worker)
IS: is a test for data movement (sort of a very large array of integer)
EP: is a strong test for floating-point performance (no communication)
FT: is a test for both communication speed (few MB messages are sent) and computation performance, thanks to a Fourier Transformation.

All these benchmarks comes with different size of problems : S,W,A,B,C and D
S is the easiest while D is the hardest. A reasonable size to use is C.

Example of usage:
$ cd ./compile
$ ./build compile.benchmarks
$ cd ../scripts/unix/benchmarks
$ ./nas_benchmarks.sh CG A 4


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@8789 28e8926c-6b08-0410-baaa-805c5e19b8d6