PICKED FROM TRUNK 16522
work related to SCHEDULING-736 and SCHEDULING-735 but fixes none of them.

Improved the IO threads to avoid readLine() deadlocks (this can occur if the remote process starts to write a line and waits before sending a LF (or CR) ).

In Scilab extention, now sending the name of the scilab function to use. Improved how remote scilab function files are handled.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/branches/Scheduling_2.1.x@16556 28e8926c-6b08-0410-baaa-805c5e19b8d6