SCHEDULING-856 Nodes topology feature in the resource manager.
Description:

the support of nodes topology consist of two steps

1. Topology information discovery which includes for now only distances between hosts. Distances are obtained during the process of adding nodes to the resource manager and supposed to be symmetrical. As this feature in the first place is targeted for native tasks the pinging process is oriented to hosts availabilities not nodes availability though Proactive (InetAddress.isReacheble() is used).

2. Selection mechanism extension. ResourceManager.getAtMostNodes(...) will accept a topology descriptor saying which set of nodes are required for computations. The following descriptors are available

ArbitraryTopologyDescriptor - no constraint on node location
BestProximityDescriptor - the set of closest nodes
ThresholdProximityDescriptor - the set of nodes within a threshold proximity
SingleHostDescriptor - the set of nodes on a single host
SingleHostExclusiveDescriptor (not implemented) - the set of nodes of a single host exclusively (the host is reserved for the user)
MultipleHostsExclusiveDescriptor (not implemented) - the set of nodes of several hosts exclusively (hosts are reserved for the user)

All descriptors allow to specify set of pivot nodes which will be used as start point for all selection algorithms.
If selection scripts are used in the "getAtMostNodes" request the node pool is first filtered out according to scripts results, then the topology information of matched nodes is considered.
Topology view have been added to the RM GUI.

Implementation details:

- In order to find nodes with best proximity 2 algorithms are implemented and used
1. Agglomerative hierarchical clustering (http://en.wikipedia.org/wiki/Cluster_analysis#Agglomerative_hierarchical_clustering)
This algorithm is quick but may not give the complete node set when there are a lot of non-connected once in the graph.
2. In such case the selection process switch to optimal clique search algorithm
http://en.wikipedia.org/wiki/Clique_problem#Cliques_of_fixed_size
which is actually a brute force algorithm with many optimizations.

- Threshold proximity selection is implemented using clique search algorithm but in this case there is no need to find the optimal clique.

- Distances between nodes on the same are zeros, so single host nodes could be found as threshold proximity nodes with threshold value equals zero.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@17569 28e8926c-6b08-0410-baaa-805c5e19b8d6