BUILD-2 Remove unnecessary jars stored in svn, improve dependencies management in the ProActive projects
It was decided that each project depending on ProActive Programming should have ant target copying required jars from the Programming. In this change it was implemented for the scheduling project, target 'copy.dependencies' was added in the build script, for now it copies just ProActive.jar and ProActive_utils.jar (and it requires build property 'programming.project.dir' to be set).
  

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21411 28e8926c-6b08-0410-baaa-805c5e19b8d6