BUILD-2 Remove unnecessary jars stored in svn, improve dependencies management in the ProActive projects
Preparing for removing content of the 'lib/ProActive' in the scheduling project:

Content of the folder 'scheduling/lib/ProActive' is produced as result of the ProActive Programming compilation, it is just a copy of the 'Programming/dist/lib', so it isn't necessary to store in the svn all these copied files. Scheduling project has special ant target 'copy.dependencies', I modified it to copy all binaries from the Programming. Information about additional step required to compile project was added into the project's README file.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21432 28e8926c-6b08-0410-baaa-805c5e19b8d6