SCHEDULING-1512 : MapReduce does not work on Windows
Problem: PAMapReduceJob.changeHadoopPath() does not work on
Windows. One of the reasons is that it tries to modify the path using
the value of "user.dir" property, and those 2 use different kinds of
slashes, so it fails.

Solution: use the value of
FileSystem.get(configuration).getWorkingDirectory() instead of
"user.dir" property. This is more correct anyway because what we are
trying to do is undo the effect of FileInputFormat#addInputPath(), and
that one uses getWorkingDirectory(), not "user.dir". Output path does
not need that kind of processing, so I removed it.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21472 28e8926c-6b08-0410-baaa-805c5e19b8d6