SCHEDULING-1520 Finished jobs were displayed as pending in the studio RCP
Problem:
Sometimes studio RCP and Web GUI displays finished jobs in the Pending list, but status for these jobs is displayed as Finished.

Fix:
One potential reason of this bug was found: scheduler's log shows that sometimes scheduler submits two jobs with the same JobId (and it was observed for the finished jobs which are displayed as pending in GUI). Jobs with the same JobId can put scheduler database in inconsistent state and cause behavior observed in SCHEDULING-1520.   
New Job to submit is generated in the SchedulerFrontend.submit and then passed to the SchedulerCore.submit. It seems code generating JobId is correct and can't generate two equal JobIds, the only suspicious place I found is mechanism passing generated job from SchedulerFrontend active object to the SchedulerCore active object. To pass job they use special object InternalJobWrapper shared between two threads, InternalJobWrapper's methods get/set are not synchronized so according to the java memory model there is no guarantee that SchedulerCore's thread will see the latest value stored in the InternalJobWrapper by the SchedulerFrontend. As result when SchedulerCore.submit is called it can see stale reference to the previous submitted job and submit it again. To fix this issue I added synchronized to the InternalJobWrapper's get/set methods.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21517 28e8926c-6b08-0410-baaa-805c5e19b8d6