SCHEDULING-1526 Error during native task starting (NPE in NativeExecutable.buildEnvironmentVariables)
Problem:
Sometimes native tasks fails to start because of NPE thrown from NativeExecutable.buildEnvironmentVariables

Fix:
This NPE can be thrown if scheduling-related system properties aren't set when task starts execution. These system properties are set in the TaskLauncher's constructor and cleared only when task completes normally or when it is terminated. It seems only possible case when NPE can be thrown is when some task is terminating at the same time when another task is starting. In the current scheduler implementation when task is killed then there is no guarantee that method NativeTaskLauncher.doTask for killed task had finished before execution of NativeTaskLauncher.doTask for another task started in the same VM. And it is possible that finishing task can clear system properties just before starting task tries to access them. 
Implementing synchronization for killed tasks can impact scheduler performance and probably it isn't really needed since this bug is quite rare, so I just implemented fix for this particular problem with system properties clearing: when task is killed then method 'unsetEnv' is called during termination and there is no need to call it again when method doTask finishes.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21526 28e8926c-6b08-0410-baaa-805c5e19b8d6