SCHEDULING-1552 ConcurrentModificationException when scheduler event is sent to the client
Problem:
Sometimes sending scheduler event to the client fails with exception and after this client is disconnected.

Fix:
When SchedulerFrontend receives JobState instance from SchedulerCore it passes it to the threads sending events to clients and stores this JobState. When job is updated SchedulerFrontend updates stored JobState instance, and it is possible that at the moment when thread sending event uses JobState instance SchedulerFrontend's thread modifies this instance and thread sending event can get ConcurrentModificationException (also this can lead to the situation when client will get event with inconsistent JobState). To fix it I changed SchedulerFrontend to store and update copy of JobState received from SchedulerCore.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21593 28e8926c-6b08-0410-baaa-805c5e19b8d6