SCHEDULING-676  be able to change the RM even if the currently used one is not dead
To test this feature it should be possible to start from the test two
RMs. To implement it RMTHelper was modified to not use static
variables/method, so that it is possible to have two instances of
RMTHelper in one test. Since its methods were changed to be non-static
all existing tests were updated: instead of calling
'RMTHelper.someMethod' tests call
'RMTHelper.getDefaultInstance().someMethod'.