SCHEDULING-676  be able to change the RM even if the currently used one is not dead
Logic of class managing scheduler's connections to the Resource manager
(RMProxiesManager) and logic of proxies was modified so that user
proxies keep reference to the RM which was used to request nodes so that
even after scheduler was linked to another RM stored connection can be
used to properly release nodes. 
Added new tests for implemented changes (TestRMProxy and
TestRMProxyRebind check logic of RMProxiesManager without scheduler,
TestLinkAnotherRM is more complex test starting scheduler with two RMs).