https://bugs.activeeon.com/browse/SCHEDULING-1609
ClientJobState is incomplete

The problem was that the class
org.ow2.proactive.scheduler.job.ClientJobState was meant to be a lighter view of the JiobState object (internal view of the Scheduler)

The ClientJobState overrides the JobState but, when constructed, the copy constructor did not copy all the fields into the new object.
I have copied part of the fields in the ClientJobState and override the getter methods for fields that are not available and throwed Runtime exceptions.

The same for the ClientTaskState.