SCHEDULING-1695 If client disconnects during getNodes request do not mark nodes as busy.
Client can disconnect ifrom the RN at any moment and it should
be consistent with ResourceManager.getNodes request. If in the
process of the nodes selection client is disconnected all the
nodes already marked as busy for this client will be freed by
the client pinger but we should stop setting other nodes as
busy.

Change-Id: I7e09d67848a46ff491726d2600bb171967cf7abb