SCHEDULING-1726 Automatic selection of network interface depending on the url of the resource manager.
If a node has several network interfaces proactive might choose an incorrect one so that the node
will not be visible from the resource manager. It's better to choose it dynamically based on the
rm's url.

If proactive properties (netmask or interface name) are defined they must cancel this automatic
detection.

The same behavior was added to the rm/scheduler command line clients.

Change-Id: I2a9da268de3efa33c96027c5d90202f51c4816fc