SCHEDULING-1743 Killing job freezes the scheduler
Problem:
SchedulerCore inside the 'endJob' method (this method is called when job
is killed) executes synchronous remote call 'TaskLauncher.terminate'
from the SchedulerCore thread, so if this call hangs then Scheduler
completely hangs. (note: 'endJob' is not the only place when remote call
is executed from the SchedulerCore thread, but hanging during job
killing is often reported by the users).

Fix:
When user kills single task then remote call 'TaskLauncher.terminate' is
executed from the special thread pool. I changed method 'endJob' to do
remote calls from the same thread pool.  After this change test
TestProcessTreeKiller started to fail since it expected that task
processes are killed synchronously, I changed test to wait some time
before checking that task processes are killed.  

Also in this change I fixed some issues in the
'restartTaskOnNodeFailure' method (this method is called if pinger
thread or thread executing TaskLauncher.doTask detects node failure). It
is possible that method is called after job or task was terminated, in
this case 'currentlyRunningTasks' map doesn't contain data for restarted
task, this could cause NPE (line 815 before this change). Also in case
if task was terminated this method shouldn't try to release NodeSet
since it should be already released. 

To test fix it is necessary to simulate node hanging, it was implemented
with the usage of the java debug interface (JDI). If node jvm is started
with special flag then it is possible to connect to this jvm remotely
using JDI API and suspend it. I created simple class 'DebugHelper' which
provides facilities for debugger connection, jvm suspending and
resuming. New test 'TestKillWhenNodeHangs' uses DebugHelper to suspend
node runnign task and checks that method 'killJob' and 'killTask' don't
hang when node is suspended.