SCHEDULING-1737 - Zombie Native Processes are not killed automatically at the end of a task
I used the ProcessTreeKiller as it was used before in the killing of a NativeExecutable. The PTK creates an environment variable cookie which is passed when creating a process via a ProcessBuilder. When killing a process tree, it will look for all those processes containing this cookie in their environment. But in order to make this work for any kind of Executable, I had to find a way to modify the environment variables of the current JVM. This was done via a small class called Environment which does some jna calls in order to call the C library (libc.so on linux and msvcrt.dll on windows).

This patch includes several improvements and modifications :
- add the Environment.java util class to modify the current JVM environment
- add the OperatingSystem and OperatingSystem family util class which has a quite extensive list of checks for Operating Systems
- add a logger to the ProcessTreeKiller class with some debug info
- modify the TaskLauncher, JavaTaskLauncher, ForkedJava... classes to make sure that the PTK is called when killing a task or after normal termination
- edited the ProcessTreeKiller test to test both killing and normal termination cleanup for JavaExecutable, ForkedJavaExecutable and NativeExecutable

Change-Id: I25f356a15f6a671427b85e2b8b612f928a3bdfd7

a vanilla commit to test the connection with Hudson for scheduling

Change-Id: I5ab8cab566566d49d8c1f8d5cd0fe6349a571b8d

a vanilla commit to test the connection with Hudson for scheduling

Change-Id: I25f356a15f6a671427b85e2b8b612f928a3bdfd7

SCHEDULING-1737 - Zombie Native Processes are not killed automatically at the end of a task

Change-Id: I25f356a15f6a671427b85e2b8b612f928a3bdfd7

SCHEDULING-1737 - Zombie Native Processes are not killed automatically at the end of a task

Improved the OperatingSystem class to handle unknown Windows OS. Added an explicit Exception when the OS cannot be found

Change-Id: I25f356a15f6a671427b85e2b8b612f928a3bdfd7

SCHEDULING-1737 - Zombie Native Processes are not killed automatically at the end of a task

changed some exception printing to debug mode

Change-Id: I25f356a15f6a671427b85e2b8b612f928a3bdfd7

SCHEDULING-1737 - Zombie Native Processes are not killed automatically at the end of a task

- rebuilt the TestSleeep.exe executable to make it independant of any dll and able to run on any Windows System
- minor changes in the test which now iterates 4 times.

Change-Id: I25f356a15f6a671427b85e2b8b612f928a3bdfd7