SCHEDULING-1756 Storing rm's node sources in the data base.
Now all the node sources are stores in the data base
and recreated after the rm restart.

There is a way to drop node sources from the data base without
loosing the node usage history. It can be done by specifying
-Dpa.rm.db.hibernate.dropdb.nodesources=true property or
launching rm-start-clean-ns scripts.

The data base manager was replaced by the similar one used
in the scheduler.

The execution of sql native queries was removed in order to avoid
compatibility issues with databases. As a result the UNION sql statement
was replaces by 2 corresponding statements as it's not supported
by hibernate.

Change-Id: If96c3c407c562e8ae525213d31665f7d546582d0