SCHEDULING-1767 Provide API to get partial scheduler state
Added new method 'loadJobs' in the Scheduler interface, it allows to get
jobs subset with pagination support. Also various filtering criteria can
be specified (JobFilterCriteria class):
- return myJobsOnly
- return pending jobs
- return running jobs
- return finished jobs
Logic for loading jobs is implemented in the SchedulerDBManager (for
paging and filtering  Hibernate API is used). SchedulerFrontend just
executes the same security checks as for 'getState' and calls
'SchedulerDBManagerloadJobs'.  

Rest API will need information about job owner, but it wasn't available
in the JobInfo interfaces so I added it.