PORTAL-263 Rest Documentation should have a header
Include a header in all pages of the Rest Documentation
The header shows the logos and the product name.

The JAX-* doclet tool has been updated to a SNAPSHOT version to include
this pull request https://github.com/FroMage/jax-doclets/pull/55
(to allow for relative paths for the logos)

The REST API Version is also set in the webpage at build time using filtering.

Change-Id: Ic76b3249e12ed3cfb80e4838ad0abd377b6b8f85