SCHEDULING-1774 Expose usage data for accounting
Add new method in the Scheduler to retrieve user data for a given period.
The methods related to accounting/usage are extracted in a new interface.
Scheduler implements it.

New DTOs describe the usage data structure and they are also used by the rest API.

Security config updated with the new method.

Change-Id: I4afb42376bbfc9c2ad00cf18f4917c8555beca59