PORTAL-260 Livelogs are tricky to use
Use the livelogs available in the user's session which means the novnc URL
is not reusable if livelogs are activated for another job.
This way we avoid attaching a new log listener to the scheduler every time
a novnc page is loaded.

Cleanup unused/dead code in live logs management.

Change-Id: I2c7123d0d179eb1a1c345e5a6233e3c7775f19f9