SCHEDULING-1809 Script tasks
Implement a new type of task: script task.
Very similar to pre/post/clean/selection scripts expect it is a task as a
native or java task.

Job Schema has been updated to 3.2 to reflect the new tag <scriptExecutable>
To ease the transition between schema versions, the schema checker is now
matching the job schema version in xmlns attribute with the schema it knows
about (starting from 3.1). We should be able to send 3.1, 3.2 or dev jobs.

The script tasks is ran using JSR 223 and the ScriptEngine. Several bindings
are available:
- result to set the task's result inside the script
- results to get the dependent tasks results inside the script
- localspace, global, user, input, output to get the dataspace file objets
  inside the script
- progress to set the task's progress inside the script
- other pas properties can be retrivied as system properties in the script
  (job/task ids, replication/looping index...)

Change-Id: I73ac00d06733ca9b273e94dd024a694e7fe7e50d