SCHEDULING-1851 Switch to jline2
Use jline2 included with jruby.jar, remove the patched jline.jar.

The patched FileNameCompletor from the removed jar is included -- the
main difference from upstream is the use of "/" in the matchFiles
method even on windows, which allows to do completion using forward
slashes.

There is no ClassNameCompletor in jline2, so i also copied it and its
dependency SimpleCompletor.

Change-Id: Icfdce320700169f45e752d71a545562650f97b7e