SCHEDULING-1912 CLI: use default credentials
When using command line tools such as rm-client, scheduler-client and rm-start-node,
they will require the user to input the credentials, either a file or a login/password.

Most of the time we use the default credentials so it would be nice to use them if they
can be found. Of course if the user provides a credentials file or a login it should be
used instead of the default credentials.

rm user will be used for rm-client and rm-start-node
scheduler user will be used for scheduler-client

The method from PASchedulerProperties#getValueAsStringOrNull... has been added to
PAResourceManagerProperties.

Change-Id: Id1b6fe1ec0cd81849bd28a489df6480f9624866a