SCHEDULING-1950 RMStart/SchedulerStarter: simplify execution of starters
Remove the need for explicit parameters, RM_HOME, SCHEDULER_HOME will be the
current folder if not specified via system properties.
Security policy will be picked up in config folder if not specified.
Log4j config will be picked up in config folder if not specified.

Extract the properties file loading logic to a specific class for RM and Scheduler.

Change-Id: I734d29823033a4e0eb6997d65699e5811f0c8003