SCHEDULING-2002: Java8 Nashorn replaces Rhino (JS)
With the introduction of Nashorn, the JS syntax for Java interoperability
changes. Either we use full class name or we include a JS script to keep the
compatibility.

The security manager is not properly handled with Java8 so some of the scripts
used for testing have been changed to Groovy

Change-Id: I6e047697fd9cbf2ad28a7060b8f53f5d66cce70e