SCHEDULING-2022 Fixed NullPointerException in scheduler functional tests.
The problems was related to the fact that the last finished task in a job changes job's state to finished.

SchedulerFrontendState.java:
    public synchronized void taskStateUpdated(String owner, NotificationData<TaskInfo> notification) {
        jobsMap.get(notification.getData().getJobId()).update(notification.getData());

Then client sees the job as finished before running_to_finished event is actually delivered to him and job "finished"
flag is set in SchedulerFrontendState.jobs list.

Scheduler.getJobResults() returns null at this moment.

Change-Id: Iee3c7fe74e64b1630d638f8943a400fae5a829e9