PORTAL-395 REST-based publisher/subscriber mechanism for Scheduler Events
ProActive Scheduler (REST) Eventing mechanism provides the ability for its users to subscribe and receive various scheduler events such as ‘change of scheduler state’, ‘job submission’, ‘change of the state of submitted job’ etc.

This implementation is based WebSocket tcp based HTTP like protocol which supports asynchronous, full duplex communication. Atmosphere 2.0 framework provides a higher level API for dealing with WebSockets in a  RESTEasy + Jetty environment.

REST Scheduler exposes ‘/eventing’ as the eventing URL which accepts HTTP verbs, GET and POST. HTTP GET requests the server an upgrade of the connection (from HTTP to WebSocket) where as HTTP POST contains the details for the actual subscription. Once the subscription is established, the connection remains open until the explicitly or implicitly close by either party.

For each subscription request, REST scheduler creates a SchedulerEventBroadcaseter instance and associates it with the subscription. Each instance implements the SchedulerEventListener interface and  REST scheduler registers them as a event listeners with back-end scheduler with appropriate parameters in accordance with the subscription request. Upon receiving a scheduler event, the SchedulerEventBroadcaster instance transmit it back to the associated client.

REST Client uses the Atmosphere Client framework to establish WebSocket based connection with the server, send the subscription request and receive the corresponding events. SchedulerEventReceiver contains the encoder/decoder for message submission and reception, and a handler to relay those events the actual listener instance.

Change-Id: I86cf5aca8e3b86ec360b9ae4209ed728dfae23b4