SCHEDULING-2037 Use same directory for scratch/local space and working dir for forked tasks
The idea is to create an output space from the forker that can be shared with the forkee.
The forkee then uses this space as the scratch space.

Functional test that copies files using relative paths.

Scratch dir property has been removed from forked launcher as the scratch space will be picked
up from the forked.

Change-Id: Ic6c5f37224f3d9d8a234b74cc84cc9aa18f06614