SCHEDULING-2063 classpath dist/lib/* contains conflicting versions of jline
Jython and JRuby come with jline embedded in the JARs. We use a more recent
version for REST CLI.

Jython and JRuby have been rebuilt with a different package name for jline.

One issue might be that JLine won't load on Windows for Jython, but we don't
use the interactive mode when running as script engines.

The changes for JRuby and Jython can be found here:
- http://gitorious.ow2.org/ow2-proactive/jython/commits/v2.5.4rc1-jline
- http://gitorious.ow2.org/ow2-proactive/jruby/commits/1.7.4-jline

To build Jython:
- run ant jar-complete
- use dist/jython.jar

To build JRuby:
- run ant jar-dist
- use lib/jruby.jar

The JARs have been uploaded manually to http://repo.activeeon.com/content/repositories/thirdparty/
- with Nexus web interface, select the repository
- use artifact upload tab
- set GAV parameters manually
  - org.python:jython-jline:2.5.4-rc1
  - org.jruby:jruby-jline:1.7.4

Change-Id: Ib2e579298cb2d12d2c8094d2fa2e2ab4039f6f16