SCHEDULING-2051 Recurring DB errors when restarting the scheduler
Do not restore "LocalNodes" node source when starting the Scehduler.
It will be recreated by SchedulerStarter with the correct number of nodes.
It means that any node source with the name "LocalNodes" will be deleted on Scheduler startup!

This way we dont try to create a NS that has been recovered upon Scheduler restart.

The timer for the RestartDownNodes policy needs to be canceled when stopping it.

Change-Id: I792d5afa80a04ff9b0ad649c114c9f038a3a0a70