SCHEDULING-2079 fix default PNP port and disable PAMR as additional protocol
PNP default port is now 64738. Separate config file for network now exists for
server and nodes because the port needs to be fix only for server.

PAMR disabled because of the error message that can be thrown if no router exists
and because it won't be often used for people trying the product.

Change-Id: Iddd72e61d14dbb055a57893bfc3b8230c8e9a6ff