SCHEDULING-2112 rest-client jar is missing in distribution
The rest-client JAR is now included in dist/lib and will be packaged in distribution.

Also removed the dependency to rest-cli in rest-client to avoid a dependency on
programming-core (which was the point of rest-client).

HttpUtility has been duplicated because I could not find a module to put it into without
adding extra dependencies (http apache).

It is also possible to build a zip of rest-client using gradle distZip in rest/rest-client,
it contains the JAR and its dependencies and can be used for clients to integrate with us.

Change-Id: I400d48bb9252115d708b6d646ea57bda7e1a97b0