SCHEDULING-2106 Deadlock in server when submitting and processing a lot of jobs
See the BUG report for a more detailed explanation but the basic idea is that
it can happen that 5 threads updating tasks when they finish deadlock.

Because they need to acquire more connections once they are locked to unlock
and because we don't timeout on connection acquisition it can lock forever.

Setting a bigger size for the connection pool helps. But similar deadlock can
still happen, they will be handled by with transaction timeouts and retried.

Maybe the pool max size will be too small for large installation, do not hesitate
to tune it.

Change-Id: I09c7bc794603ef2652f2a06abcfa447491fb3557