SCHEDULING-2201 Discovery service for nodes
The Scheduler when started, will listen to broadcast queries (by default on 64739)
to allow Nodes to be started without a RM URL.
The Scheduler will reply to Node's queries with the RM URL.

Nodes can be started without -r parameter and will therefore use discovery.
After a timeout (configurable), the node will exit.

Discovery can be disabled and the port configured.

Change-Id: I29ec730ce8aaedcd349be1fd30034d3433f2bfb1