SCHEDULING-2225 Job recovery can fail for skipped tasks
In case of skipped tasks, the execution order was not computed correctly
because they all have the same finishedTime. It was causing the recovery
to fail.

Execution also relies on task dependencies now to avoid this.

If recovery fails for a job, the job will be marked as CANCELED and will not
be candidate for execution. Errors will be logged. The Scheduler will still
start with eventually some jobs canceled with tasks in inconsistent states
(but not running).

Change-Id: I89d607bfcdb1f162d5c90b37eb27e0db92526d2e