SCHEDULING-2236 Network failure when finishing tasks can lead to task running for ever
In case the Scheduler cannot be notified of the task termination, the TaskLauncher
will terminate on its own. The expected behavior is then that task pinging on the
Scheduler side will fails, letting the Scheduler know about the failed task.

In case of network failure, the TaskResultCallback AO is now terminated even though
sending the task result failed.

Change-Id: I69d94d68fdd15ce51bed50fbe4053b69395465f3

Conflicts:
	scheduler/scheduler-server/src/main/java/org/ow2/proactive/scheduler/core/SchedulingService.java