SCHEDULING-2250: Use active detection mechanism for scheduling tasks
The purpose of the patch is to improve the scheduler responsiveness for short
computations submitted in batches. To achieve this aim a synchronization based
on wait/notify is used. A timeout that wakes up the scheduling thread
periodically is still required for policies or nodes that are selected based on
time constraints.

Change-Id: I2cd295f0e614e4eab7e8f433ac9c4b9242fe2428