SCHEDULING-2257 Nodes can stay busy forever when a task fails to start
It could happen that when the nodes were released, the task was already restarted
on new nodes. Then the wrong nodes would be released and the nodes that was busy
for the failed task would stay busy forever.

The nodes running the tasks are stored along the running task data to make sure
the right nodes are released.

Change-Id: Ie44829e473c2888c2d9365643f9cf7dc13e21954