SCHEDULING-2265 Variable propagation broken when parent task exits with walltime
When updating variables with values from parent tasks, simply add them to the variables
from the job instead of replacing them. Otherwise if the variables from the parent task
are null (in case of error), the variables map would be empty, ie we loose the variables
defined in the job.

Change-Id: I375dad404cb546e03ccfe47ef2973d9a5af65846