Allow execution of server side workflows
This commit enable the Scheduler REST API to be able to create jobs
using workflows that are retrieved from HTTP servers.
Also makes the Studio REST API expose workflow and template contents
(provided valid credentials).
Consequently, now it is possible to create a workflow (or template)
using the Studio, get its URL, and create with it (and additional
variables if needed) a new job resource on the Scheduler REST API. It
will automatically perform authentication and downloading of the
workflow, and then submission.