Make sure the task launcher terminates even when not being called back
When receiving the task result, the Scheduler was in charge of calling terminate to
end the task launcher. This could very well fail leaving the task launcher running forever.

Now the task launcher exits on its own once the task result has been sent to the Scheduler.

terminate on the Task Launcher is still used to kill a running task