Avoid database deadlock upon job termination/job creation
This deadlock behavior was observed during manual stress tests of the Scheduler.
It seems related to #2265 however the synchronized methods in SchedulerDBManager
are still required.

The task executable (i.e a script) is now stored inside TaskData, similar to
pre/post/clean scripts. ScriptTaskData has then been removed and it seems to help
preventing the database deadlock that was observed when submitting and scheduling
a lot of tasks.

Pre/post/executable/clean scripts are mapped with a OneToOne relation with the task
whereas selection scripts are mapped with a OneToMany relation. Since the removal of
scripts has been simplified (to avoid the in query, that breaks with a lot of tasks),
selection scripts needs to be stored in a different table. Otherwise all scripts linked
to a task in ScriptData are considered as selection scripts for this task.

TaskData needs to be updated as well when removing scripts (removeJobScripts) otherwise
recovery fails (observed with TestIfTaskRestore).