Synchronize access to methods that perform job insertion and deletion in the DB
This patch fixes #2265.

Tests which have been written but also the fix are for the issue which has been
discovered. However, it seems possible that a similar issue can occur with other
methods from SchedulerDBManager, especially those that involve writes in the DB.

Extra testing is let for future work.