NodeSource threadpool could get stuck when adding/removing node sources
Functional tests showed that if the last node source is removed, then
the threadpool will be asked to shutdown, however new requests, i.e the
creation of a new node source, could happen in the middle, causing the
shutting down of the threadpool to hang and the new nodes acquisition
to fail.

Here is the observed sequence of actions:
1) last node source is removed
2) node source threadpool is shutdown
3) while threadpool is shutting down, a new node source is created
4) the new node source uses the threadpool for node acquisition (acquireNode) and blocks until the node is acquired
5) when started the node calls addNode on RMCoren that submits a NodeLocator task to the node source threadpool
6) the node locator task cannot be submitted, because the node source threadpool lock (synchronized methods) is owned by shutdown
7) shutdown of node source threadpool cannot end because acquireNode is still running (it will timeout eventually)
8) acquireNode times out, nodes are marked as lost, the node source threadpool shutdown finishes
9) the node source threadpool is recreated, NodeLocator can run
10) however nodes are not expected anymore in the node source, node acquisition fails

My fix is simply to not shutdown the node threadpool when there is no node sources.
In real life, there is little chance that there is 0 node sources. I guess the idea
was to shutdown properly the thread pool upon RM shutdown, however with the previous implementation,
if the thread pool is busy during RM shutdown, it would still block.

If a never ending operation is executed in the node source threadpool then it might block
the RM shutdown.