Fix race condition between down node detection and task rescheduling
If a node is marked as down, then the task restarted, the node will be released.
When being released the node will be marked as free again (and eventually then marked as down).
The task can be rescheduled on the same node again since it is free but will fail to start.

This change probably breaks SCHEDULING-1979 Improve handling of node disconnection
(efe0f90c3ba883b385f3cb431b1e17bfa5731c6d).

However considering that #1986 is still open, I would not consider the support
of node disconnection stable for now.

This race condition was revealed by TestTaskRestartOnNodeFailure.