Let TaskLauncher handle node shutdown
If a node shutdown happens, then it can cause the task to end and yet
to send the result (a failure since it has been interrupted) to the
Scheduler. We would rather have the Scheduler keep the task as running
and then detect on its own that the node is not there.