Avoid too long command line (classpath) on Windows for local nodes
On Windows, command line length is limited, using java.class.path caused it
to be too long. It means the IDE classpath won't be picked up sadly (i.e
we have to run gradle dist, if scheduler classes are changed).

It also means that rest smart proxy test classes need to be in the scheduler
classpath (dist/lib) when running tests.