Refactor methods and fix missing parameters
Fix constructor calls in SchedulerFrontend and JobStateTest. Refactor method names in JobState and Scheduler. Implement getTaskResultsByTag and removeJob in SchedulerClient. MAX_PAGE_SIZE is now parametered on server-side (pa.scheduler.tasks.page.size) and as a fixed value for the CLI in RestConstants (using the same value as default page size for jobs)