Remove CancelOnTaskError and replace with OnTaskError.
CancelOnTaskError is deprecated and the internal data structures are changed to not hold this property. Workflows which have this property set will be converted. The conversion is that a CancelTaskOnError is transformed into onTaskError=cancelJob.

A lot of the logic relied on the cancelJobOnError property. All logic ins now converted to use the new property but keeps the old behavior.