Fix functional test workflows, convert cancelTaskOnError to OnTaskError and fix broken tests and fix restart count bug.
The functional tests need conversion because the Workflow xsd was changed. Workflows are converted and Workflows which use the cancelJobOnError are kept and converted into the new version 3.4. So backwards compatability tests are kept and the tested characteristics are translated into the new Workflow model.
The PauseTaskOnError changes broke a few tests, those are fixed by this.
The changes in the LiveJobs changed the way how executions are counted, it happened that a restart over the REST API did not count as an execution. But functionaltests relied on that fact. So the counting behavior was restored with the possibility of introducing the "2 times counted bug". The 2 times counted bug counts execution when an error appeared two times. Nevertheless it could not be validated that the bug appears again, due to the complexity of state changes and complexity of the code in LiveJObs.java, it can't be ensured that the 2 times counted bug does not appear.