Add setter for maxNumberOfExecutionsLeft and maxNumberOfExecutionsOnFailureLeft
The set maxNumberOfExecutions setter was overriden and does set maxNumberOfExecutionsLeft and maxNumberOfExecutionsOnFailureLeft as well. As a matter of fact, the data which loaded from the database was inconsistent with the data displayed. (I am wondering why there are two layers of data representation, and why saving it in the database but then not using it?). The data inconsistency appeared between TaskData, which holds the database data and InternalTask, which holds some other data (a mixture of database values and constants).
Additional setters are introduced in order to set the number of executions left after being read from the database. So the number of executions left, read from the database, are set by setters in the InternalTask class.