Force read-only requests from a SessionWork to be executed in a transaction.
Executing a select query that involves large objects must be done in a
transaction at least for Postgres database. In the general case it is also a
good practive in order to get a consistent view of the database. Imagine that
you have a read-only SessionWork with 2 or more select queries. In a highly
concurrent scenario it could happen that data which have been read in the first
select is modified before the other selects are executed.

This patch fixes #2425.

It also includes a few enhancements:

- factorization of the code used to manage transactions by reusing the class
  TransactionHelper with SchedulerDbManager and RmDbManager.
- rename transaction related methods runWithoutTransaction/runWithTransaction
  to respectively executeReadOnlyTransaction and executeReadWriteTransaction
- remove SchedulerDbManager#setCallback method that is not referenced by other
  projects. It allows to delete the obscure DatabaseManagerExceptionHandler