Low priority tasks can be scheduled over high priority ones #2544
The InternalPolicy filter can remove high-priority tasks from the list of tasks to be scheduled, making priorities irrelevant.

When calling getRMNodes() in the scheduling loop, the set of nodes available on the resource manager at this moment can differ from the set of nodes which were available at the beginning of the loop. Accordingly, a low priority task can take this opportunity to schedule before higher priority ones

Included in this change :
- remove the internal policy filter
- add a check for priority in SchedulingTaskComparator (to ensure that high and low priority tasks are not deployed concurrently)
- update free nodes number at the beginning of the scheduling loop only. Otherwise, nodes can be released inside the loop and change the scheduling policy behavior (Fifo order, priorities).
- memorize inside the loop the expected number of nodes available
- break the loop if the actual resource manager state differs from the expected
- add various tests which check that priority is preserved and that starvation does not occur in case high-priority tasks cannot be scheduled
- add a complex test which tries to create a large number of nodes inside a scheduling loop in order to reproduce the issue
- add a TestScheduler non-fork-mode-with-restart superclass in the testsuite