Limit the maximum number of Nodes the Resource Manager accepts.
This design limits the maximum number of Nodes by rejecting the Node in the internalAddNodeToCore method. That method is called after the Node is registered inside the Resource Manager and an AddNode event is send.

When the Node limit is reached the Node is rejected and removed from the Node list. One will see an exception inside the Scheduler.log. The Node executable will just be killed from the Resource Manger, without any error message, or sometimes with a stacktrace which indicates that the node was interrupted.

This solution is different to the solution of adding the check, of how many nodes are allowed, into the addNode method. In that method the error message is forwarded and displayed in the node output. That is understandable for the user, but it introduces race conditions and more nodes than wanted can register at the Resource Manager.

Furthermore, the code is compiled with a maxNumberOfNodes = -1; That ensures that this feature is deactivated all of the times. Only if the release script explicitly changes the value inside the .class file it will active. That hinders us from using a final static, because the compiler will do optimizations inside the bytecode, which compiles the constant away and it removes the constant from any if.
As a result, a change of the variable inside the byte code has no effect on the node limitation.