Add new group and user allowed to connect/disconnect and add nodes to RMCore
The patch adds a new group called "node" along with a new user "node" that is
member of the last group. All users assigned to "node" group are allowed to
connect to RMCore, add a node in a node source and eventually disconnect.

The credential file associated to user "node" is provided by default. It aims to
replace in a future patch the file "rm.cred" included in node.jar.