Add variable recursive late bindings for task generic information
Before this change, variable patterns used in generic info were static and did not use propagated variable values.

In order to allow late binding on generic info, InternalTask keeps a map of up-to-date variables. SchedulingMethodImpl class updates this map the first time it tries to schedule a given task.

To allow recursive binding where the name of the variable used is built by another variable, the VariableSubstitutor was updated.

Additionally
- refactored InternalTask to hold a transient reference to InternalJob (simplification of APIs)
- allow recursive late bindings for variable themselves
- added a unit test for recursive replacements
- added a functional test which verifies late bindings for generic info