Skip OS Name test
What does this test test?:
It seems that the test gets os.name (System.getProperty) and takes it as a regular expression.
Then the test gets os.name (System.getProperty) again and matches that against the regular expression.

Problem:
Windows Server 2016 return Windows NT (unknown)
"(unknown)" is not interpreted as a simple string because the brackets are not escaped
Matching against the same string fails.

Solution:
Making the os.name string a literal regex pattern