Add files via upload
Add resultMetadata template.

Add files via upload

Update job.xml

improve email template - add job id in the body and remove the release version

remove unused import

remove the pompt of starting calendar service in bat file, as did in bash file

improvment on the setup script, package installation not necessary if it was already done

define and use a variable for local radicale path in the script

define and use a variable for local radicale path in the script

Update job.xml

Update job.xml

Fix SQL error on shutdown when embedded HSQLDB server is used

The default embedded database must not be closed in a shutdown hook.

Multiple shutdown hooks are defined. Some are used for instance by the
Resource Manager to remove node sources before termination. If the database is
closed before completing the last operation, then some errors will occur
(see logs).

Besides, checking the state of the Scheduler and RM by registering an event
listener or even by polling with periodic method calls will not work since
Scheduler and RM are Active Objects and these are terminated by other shutdown
hooks.

The database connection should be closed by Hibernate when the datasource is
closed. If not, the HSQLDB server will handle in the worst case the JVM shutdown
as an accidental machine failure.

Skip OS Name test

What does this test test?:
It seems that the test gets os.name (System.getProperty) and takes it as a regular expression.
Then the test gets os.name (System.getProperty) again and matches that against the regular expression.

Problem:
Windows Server 2016 return Windows NT (unknown)
"(unknown)" is not interpreted as a simple string because the brackets are not escaped
Matching against the same string fails.

Solution:
Ignoring the test for now, because I am not sure if it makes a lot of sense to check if System.getProperty("os.name") matches System.getProperty("os.name")

Skip OS Name test

What does this test test?:
It seems that the test gets os.name (System.getProperty) and takes it as a regular expression.
Then the test gets os.name (System.getProperty) again and matches that against the regular expression.

Problem:
Windows Server 2016 return Windows NT (unknown)
"(unknown)" is not interpreted as a simple string because the brackets are not escaped
Matching against the same string fails.

Solution:
Making the os.name string a literal regex pattern

improve email template - add job id in the body and remove the release version

remove unused import

remove the pompt of starting calendar service in bat file, as did in bash file

define and use a variable for local radicale path in the script

Fix SQL error on shutdown when embedded HSQLDB server is used

The default embedded database must not be closed in a shutdown hook.

Multiple shutdown hooks are defined. Some are used for instance by the
Resource Manager to remove node sources before termination. If the database is
closed before completing the last operation, then some errors will occur
(see logs).

Besides, checking the state of the Scheduler and RM by registering an event
listener or even by polling with periodic method calls will not work since
Scheduler and RM are Active Objects and these are terminated by other shutdown
hooks.

The database connection should be closed by Hibernate when the datasource is
closed. If not, the HSQLDB server will handle in the worst case the JVM shutdown
as an accidental machine failure.

Skip OS Name test

What does this test test?:
It seems that the test gets os.name (System.getProperty) and takes it as a regular expression.
Then the test gets os.name (System.getProperty) again and matches that against the regular expression.

Problem:
Windows Server 2016 return Windows NT (unknown)
"(unknown)" is not interpreted as a simple string because the brackets are not escaped
Matching against the same string fails.

Solution:
Ignoring the test for now, because I am not sure if it makes a lot of sense to check if System.getProperty("os.name") matches System.getProperty("os.name")

Skip OS Name test

What does this test test?:
It seems that the test gets os.name (System.getProperty) and takes it as a regular expression.
Then the test gets os.name (System.getProperty) again and matches that against the regular expression.

Problem:
Windows Server 2016 return Windows NT (unknown)
"(unknown)" is not interpreted as a simple string because the brackets are not escaped
Matching against the same string fails.

Solution:
Making the os.name string a literal regex pattern